move smart-tx file to utils
